+++ web/modules/custom/ai_screening/src/Drush/Commands/Commands.php
@@ -2,9 +2,13 @@
 
 namespace Drupal\ai_screening\Drush\Commands;
 
+use Consolidation\AnnotatedCommand\CommandData;
+use Consolidation\AnnotatedCommand\Hooks\HookManager;
 use Consolidation\OutputFormatters\StructuredData\RowsOfFields;
 use Drupal\Core\Entity\EntityInterface;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
+use Drupal\Core\Session\AccountSwitcherInterface;
+use Drupal\Core\Session\UserSession;
 use Drupal\ai_screening\Exception\InvalidArgumentException;
 use Drupal\ai_screening_project_track\Helper\ProjectTrackHelper;
 use Drupal\ai_screening_project_track\Helper\ProjectTrackToolHelper;
@@ -38,6 +42,7 @@ final class Commands extends DrushCommands {
     private readonly EntityTypeManagerInterface $entityTypeManager,
     private readonly ProjectTrackHelper $projectTrackHelper,
     private readonly ProjectTrackToolHelper $projectTrackToolHelper,
+    private readonly AccountSwitcherInterface $accountSwitcher,
   ) {
     $this->projectTrackStorage = $this->entityTypeManager->getStorage('project_track');
   }
@@ -107,6 +112,22 @@ final class Commands extends DrushCommands {
     }
 
     return new RowsOfFields($rows);
+  }
+
+  /**
+   * Pre command hook for pm:install.
+   */
+  #[CLI\Hook(HookManager::PRE_COMMAND_HOOK, target: '*')]
+  public function preCommand(CommandData $commandData): void {
+    $this->accountSwitcher->switchTo(new UserSession(['uid' => 1]));
+  }
+
+  /**
+   * Post command hook for pm:install.
+   */
+  #[CLI\Hook(HookManager::POST_COMMAND_HOOK, target: '*')]
+  public function postCommand($result, CommandData $commandData): void {
+    $this->accountSwitcher->switchBack();
   }
 
 }