+++ web/themes/custom/itkdev/itkdev_project_theme/tailwind.config.js
@@ -11,12 +11,25 @@ module.exports = {
     "bg-slate-300",
     "bg-yellow-700",
     "bg-red-700",
+    "bg-green-600",
+    "bg-green-700",
     "grid-cols-4",
     "order-last",
     "text-slate-300",
     "border-zinc-400",
     "border-red-600",
     "border-yellow-600",
+    "green-600",
+    "green-700",
     "border-green-600",
   ],
+  theme: {
+    colors: {
+      'green': {
+        DEFAULT: '#047857',
+        600: '#047857',
+        700: '#064E3B',
+      },
+    },
+  }
 };