+++ CHANGELOG.md
@@ -6,8 +6,21 @@ The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
+
+- [PR-80](https://github.com/itk-dev/ai-screening/pull/80)
+  - Added static select webform element
 - [PR-79](https://github.com/itk-dev/ai-screening/pull/79)
   - Added back buttons on several pages
+- [PR-77](https://github.com/itk-dev/ai-screening/pull/77)
+  - Added translation functionality
+  - Added .po files
+- [PR-76](https://github.com/itk-dev/ai-screening/pull/76)
+  - Align the green color
+  - Add default colors to base tailwind config
+  - Change description length in views
+  - Fix table cells
+- [PR-75](https://github.com/itk-dev/ai-screening/pull/75)
+  - Added configurable report types
 - [PR-73](https://github.com/itk-dev/ai-screening/pull/73)
   - Added colors to graphs
   - Added possibility to display graphs of multiple projects